backend/requirements.txt
Django==5.0.2
pre-commit==3.5.0
black==24.2.0
djangorestframework==3.14.0
neo4j==5.18.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
gunicorn==20.1.0
jwt==1.3.1
PyJWT==2.4.0
